requirements/dev.txt
uvicorn~=0.32.1
fastapi-decorators~=1.0.6
python-dotenv~=1.0.1
alembic~=1.14.1
torch==2.6.0
flask==3.0.3
gunicorn==23.0.0
ultralytics==8.3.3
pillow==11.0.0
pandas==2.2.3
opencv-python==4.10.0.84
openvino
openpyxl==3.1.5
torchvision
gspread==6.1.4
gspread-dataframe==4.0.0
fastapi
starlette
sqlalchemy
uvicorn~=0.32.1
fastapi-decorators~=1.0.6
httpx
python-dotenv~=1.0.1
razorpay
setuptools
cryptography
aiosqlite
pyjwt~=2.10.1
alembic~=1.14.1
pandas~=2.2.3
async-lru~=2.0.5
kiteconnect~=5.0.1
weasyprint
boto3
pillow
